+++ src/Client/ClientCore/Network/Infrastructure/ClientNetworkInfrastructureInstaller.cs
@@ -31,5 +31,6 @@ public sealed class ClientNetworkInfrastructureInstaller : IWindsorInstaller
         container.Register(Component.For<AuthenticationClient>().LifestyleSingleton());
         container.Register(Component.For<PlayerManagementClient>().LifestyleSingleton());
         container.Register(Component.For<RegistrationClient>().LifestyleSingleton());
+        container.Register(Component.For<WorldSegmentDataClient>().LifestyleSingleton());
     }
 }