+++ src/Common/NetworkCore/Network/Pipeline/Outbound/DeliveryMethodOutboundPipelineStage.cs
@@ -32,7 +32,9 @@ public class DeliveryMethodOutboundPipelineStage : IOutboundPipelineStage
     private readonly Dictionary<EventId, DeliveryMethod> methodMap = new()
     {
         { EventId.Core_Ping_Ping, DeliveryMethod.Unreliable },
-        { EventId.Core_Ping_Pong, DeliveryMethod.Unreliable }
+        { EventId.Core_Ping_Pong, DeliveryMethod.Unreliable },
+        { EventId.Core_WorldManagement_Subscribe, DeliveryMethod.ReliableUnordered },
+        { EventId.Core_WorldManagement_Unsubscribe, DeliveryMethod.ReliableUnordered }
     };
 
     public ILogger Logger { private get; set; } = NullLogger.Instance;